добавляет модульную организацию кода
